+++ tests/Jroc.Test262.Tests/built-ins/Number/ExecutionTests.StringOctalLiteral.cs
@@ -1,0 +1,8 @@
+namespace Jroc.Test262.Tests.built_ins.Number;
+
+public partial class ExecutionTests
+{
+    [Fact(DisplayName = "string-octal-literal.js")]
+    public Task string_octal_literal()
+        => ExecutionTestFromFile("string-octal-literal");
+}